requirements.txt
# Standard Python Package Tools
pip
setuptools
wheel

# DJANGO itself
#Django<3.0,>=2.2.9
Django>=4.0

# Application and WSGI Utils
Werkzeug
eventlet
gevent
greenlet
gunicorn
whitenoise

# Database
psycogreen
psycopg2-binary
SQLAlchemy

# Django Utils
django-admin-tools
django-appconf
django-appmedia
django-debug-toolbar
django-extensions
git+https://github.com/litchfield/django-safelogging.git@master#egg=django-safelogging
django-classy-tags

# Django Models, Fields, Addons
django-countries
django-durationfield
Unidecode
python-slugify
django-uuslug
django-taggit
django-cors-headers
django-reversion>=2.0
django-select2
django-mptt >= 0.9
django-polymorphic>=2.0 # old django-filer wants 1.3, but 2.0 supports Django2.0

# Rich Text Editors
django-ckeditor
django-tinymce

# Django-Filer
django-filer
#git+https://github.com/mci/django-filer.git#egg=django-filer

# DjangoCMS
django-cms
djangocms-admin-style
git+https://github.com/divio/djangocms-history#egg=djangocms-history
djangocms-attributes-field

# DjangoCMS Plugins
#aldryn-bootstrap3
git+https://github.com/russmo/aldryn-bootstrap3.git#egg=aldryn-bootstrap3
git+https://github.com/mci/djangocms-layout.git#egg=djangocms-layouttools
#djangocms-layouttools # pypi version pushed from mci github
aldryn-style
#djangocms-text-ckeditor
git+https://github.com/divio/djangocms-text-ckeditor.git#egg=djangocms_text_ckeditor # github more recent that pypi
git+https://github.com/russmo/cmsplugin-plaintext.git#egg=cmsplugin_plaintext_djangocms3
djangocms-file
djangocms-googlemap
djangocms-link
djangocms-picture
djangocms-audio
djangocms-video
djangocms-snippet
djangocms-style

# djangocms-teaser
# djangocms-inherit
# cmsplugin-filer
# git+https://github.com/russmo/cmsplugin-filer.git@develop#egg=cmsplugin-filer

# Images
Pillow
easy-thumbnails>=2.4
sorl-thumbnail

# Search
django-haystack

# CartoDB API
git+https://github.com/CartoDB/cartodb-python.git#egg=carto

# Authentication
git+https://github.com/mci/django-social-auth.git@django1.9#egg=django-social-auth
python-social-auth

# Spatial Libraries
Fiona
Shapely
django-shapes
geographiclib
git+https://github.com/sigma-geosistemas/django-shape-engine.git#egg=django_shape_engine

# Amazon Web Services
boto

# Location Services
python-geoip
python-geoip-geolite2
pygeoip
geoip2

# Automation
Fabric

# Testing
pytest

# Other Python Utils
argparse
BeautifulSoup4
cryptography
docutils
feedparser
httplib2
ipython
jupyter
Markdown
ndg-httpsclient
oauth2
path.py
pycryptodome
pyenchant
Pygments
python-memcached
setproctitle
simplejson
requests
requests_oauthlib
webencodings

# WIKI
#git+git://github.com/benjaoming/django-wiki.git#egg=wiki
wiki

# NOT USED
#hg+https://bitbucket.org/russmo/django-storages-multi#egg=django-storages
#git+git://github.com/jlopez/maxmind-geoip.git#egg=GeoIP-Python
#django-filebrowser-no-grappelli

